requirements.txt
# ==== Core =====
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0

# ==== LangChain =====
langchain==0.3.12
langchain-core==0.3.30
langchain-openai==0.2.14
langchain-community==0.3.12
langchain-text-splitters==0.3.3
langchain-experimental

# ==== OpenAI (Azure compatible) =====
openai==1.58.1

# ==== Data / Config =====
orjson==3.10.7
pydantic==2.8.2
pydantic-core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.32
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.2
zstandard==0.23.0

# ==== RAG Core =====

# Vector DBs
faiss-cpu==1.7.4
chromadb==0.5.3

# Document loaders
pypdf==4.2.0
unstructured==0.14.10
beautifulsoup4==4.12.3
nltk==3.8.1

# Retrieval boost
rank-bm25==0.2.2

# ==== Backend =====
fastapi==0.111.0
uvicorn==0.30.1

# ==== UI =====
streamlit==1.37.1

# ==== Compatibility =====
packaging>=23.2,<25
youtube-transcript-api==0.6.1